pyproject.toml
[project]
name = "live-translator"
version = "0.1.0"
description = "Transparent overlay window with live OCR and translation"
requires-python = ">=3.11"
dependencies = [
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "pillow>=10.0.0",
    "pywebview>=4.4.0",
    "mss>=9.0.0",
    "numpy>=1.24.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend"]

# =============================================================================
# RUFF CONFIGURATION
# =============================================================================
[tool.ruff]
line-length = 88
target-version = "py311"

# Exclude common directories
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
]

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E, W), isort (I), pep8-naming (N),
# pydocstyle (D), type-checking (TCH), and more
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "D",      # pydocstyle
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "C4",     # flake8-comprehensions
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "ANN201",  # Don't require return type annotation in tests
    "D",       # Don't require docstrings in tests
]

# =============================================================================
# PYRIGHT CONFIGURATION
# =============================================================================
[tool.pyright]
include = ["backend", "tests"]
exclude = [
    "**/__pycache__",
    ".venv",
    "build",
    "dist",
]

pythonVersion = "3.11"
typeCheckingMode = "strict"

# Strict type checking options
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedClass = true
reportUnusedFunction = true
reportUnusedVariable = true
reportDuplicateImport = true
reportOptionalMemberAccess = true
reportOptionalSubscript = true
reportPrivateImportUsage = true

# Additional strict checks
reportConstantRedefinition = true
reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportUnnecessaryIsInstance = true
reportUnnecessaryCast = true
reportUnnecessaryComparison = true

# =============================================================================
# PYTEST CONFIGURATION
# =============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# =============================================================================
# COVERAGE CONFIGURATION
# =============================================================================
[tool.coverage.run]
source = ["backend"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
